chore(http): move wiring to normalization
