+++ src/composition/http/normalization.ts
@@ -1,3 +1,17 @@
 import { createNormalizer, type Normalizer } from '@koala-ts/framework';
+import type { Result } from '@/app/shared/application/result';
+import {
+  createResultToHttpMapper,
+  type HttpResponsePayload,
+  type ResultHttpMapping,
+} from '@/app/shared/infrastructure/http/result-to-http';
 
 export const normalize: Normalizer = createNormalizer([]);
+
+export function mapResultToHttp<T, E extends { type: string }>(
+  result: Result<T, E>,
+  httpMapping: ResultHttpMapping<T, E>,
+): HttpResponsePayload {
+  const resultToHttpMapper = createResultToHttpMapper({ normalize });
+  return resultToHttpMapper(result, httpMapping);
+}
